LayoutTests/http/tests/eventsource/eventsource-reconnect-during-navigate-crash-expected.txt

CONSOLE MESSAGE: EventSource cannot load http://127.0.0.1:8000/eventsource/resources/reconnect.php due to access control checks.
